sdk: Apply relevant changes related to 3240221614727e7bb754de6b33397dc90a92ddee

+++ src/sdk/src/merkle.rs
@@ -54,5 +54,5 @@ pub fn merkle_add(
 }
 
 extern "C" {
-    fn merkle_add_(ptr: *const u8, len: u32) -> i32;
+    fn merkle_add_(ptr: *const u8, len: u32) -> i64;
 }
